+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/shipping_method/ShippingMethodUpdateActionBuilder.java
@@ -15,6 +15,10 @@ public class ShippingMethodUpdateActionBuilder {
         return com.commercetools.api.models.shipping_method.ShippingMethodAddShippingRateActionBuilder.of();
     }
 
+    public com.commercetools.api.models.shipping_method.ShippingMethodAddStoreActionBuilder addStoreBuilder() {
+        return com.commercetools.api.models.shipping_method.ShippingMethodAddStoreActionBuilder.of();
+    }
+
     public com.commercetools.api.models.shipping_method.ShippingMethodAddZoneActionBuilder addZoneBuilder() {
         return com.commercetools.api.models.shipping_method.ShippingMethodAddZoneActionBuilder.of();
     }
@@ -37,6 +41,10 @@ public class ShippingMethodUpdateActionBuilder {
 
     public com.commercetools.api.models.shipping_method.ShippingMethodRemoveShippingRateActionBuilder removeShippingRateBuilder() {
         return com.commercetools.api.models.shipping_method.ShippingMethodRemoveShippingRateActionBuilder.of();
+    }
+
+    public com.commercetools.api.models.shipping_method.ShippingMethodRemoveStoreActionBuilder removeStoreBuilder() {
+        return com.commercetools.api.models.shipping_method.ShippingMethodRemoveStoreActionBuilder.of();
     }
 
     public com.commercetools.api.models.shipping_method.ShippingMethodRemoveZoneActionBuilder removeZoneBuilder() {
@@ -69,6 +77,10 @@ public class ShippingMethodUpdateActionBuilder {
 
     public com.commercetools.api.models.shipping_method.ShippingMethodSetPredicateActionBuilder setPredicateBuilder() {
         return com.commercetools.api.models.shipping_method.ShippingMethodSetPredicateActionBuilder.of();
+    }
+
+    public com.commercetools.api.models.shipping_method.ShippingMethodSetStoresActionBuilder setStoresBuilder() {
+        return com.commercetools.api.models.shipping_method.ShippingMethodSetStoresActionBuilder.of();
     }
 
     /**